Check if Classified doesn't already exist
based on classified id parameter

+++ LBC Test/LBC Test/Data/CoreData/EntitiesDescriptions/CategoryDescription.swift
@@ -8,14 +8,14 @@
 import Foundation
 
 protocol CategoryProtocol {
-    var id: Int64? { get }
+    var id: Int? { get }
     var name: String? { get }
 }
 
 class CategoryDescription: CategoryProtocol, Decodable {
-    var id: Int64?
+    var id: Int?
     var name: String?
-    init(id: Int64,
+    init(id: Int,
          name: String)
     {
         self.id = id
@@ -24,7 +24,7 @@ class CategoryDescription: CategoryProtocol, Decodable {
 
     required init(from decoder: Decoder) throws {
         let c = try decoder.container(keyedBy: JSONCategory.self)
-        self.id = try c.decodeIfPresent(Int64.self, forKey: .id)
+        self.id = try c.decodeIfPresent(Int.self, forKey: .id)
         self.name = try c.decodeIfPresent(String.self, forKey: .name)
     }
 }
